+++ packages/devtools_extensions/lib/src/template/devtools_extension.dart
@@ -44,6 +44,10 @@ bool get _useSimulatedEnvironment =>
 ///
 /// A couple of use case examples include posting messages to DevTools or
 /// registering an event handler from the extension.
+/// 
+/// [extensionManager] can only be accessed below the [DevToolsExtension] widget
+/// in the widget tree, since it is initialized as part of the
+/// [DevToolsExtension]'s [initState] lifecycle method.
 ExtensionManager get extensionManager =>
     globals[ExtensionManager] as ExtensionManager;
 
@@ -51,10 +55,28 @@ ExtensionManager get extensionManager =>
 ///
 /// This manager provides sub-managers to interact with isolates, service
 /// extensions, etc.
+/// 
+/// [serviceManager] can only be accessed below the [DevToolsExtension] widget
+/// in the widget tree, since it is initialized as part of the
+/// [DevToolsExtension]'s [initState] lifecycle method.
 ServiceManager get serviceManager => globals[ServiceManager] as ServiceManager;
 
-/// A wrapper widget that initializes the [extensionManager] and establishes a
-/// connection with DevTools for this extension to interact over.
+/// A wrapper widget that performs initialization for a DevTools extension.
+/// 
+/// This widget is required to be at the root (or very close to the root) of
+/// your DevTools extension Flutter web app. The content of your DevTools
+/// extension should be defined by [child].
+/// 
+/// This wrapper:
+///  * initializes the [extensionManager] and [serviceManager] globals.
+///  * initializes the [extensionManager] with the VM service connection from
+///    DevTools when[requiresRunningApplication] is true.
+///  * establishes a connection with DevTools for this extension to interact
+///    over.
+/// 
+/// Any use of the [extensionManager], [serviceManager], or [ideTheme] globals
+/// must occur below the [DevToolsExtension] widget in the widget tree (i.e. at
+/// the level of [child] or below).
 class DevToolsExtension extends StatefulWidget {
   const DevToolsExtension({
     super.key,